Cambiato il tipo da int a float
Ho effettuato un cambio del tipo (da int a float) a vari variabili e a vari parametri formali dei metodi della classe Calculator

+++ progetto2509/src/main/java/unibg/it/progetto2509/Calculator.java
@@ -1,33 +1,33 @@
 package unibg.it.progetto2509;
 
 public class Calculator {
-	int x;
-	int y;
+	float x;
+	float y;
 	
-	public Calculator(int x, int y) {
+	public Calculator(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	void prod(int x, int y) {
-		int result;
+	void prod(float x, float y) {
+		float result;
 		result=x*y;
 		System.out.println("il prodotto: "+result);
 	}
 	
-	void diff(int x, int y) {
-		int result;
+	void diff(float x, float y) {
+		float result;
 		result=x-y;
 		System.out.println("differenza: "+result);
 	}
 	
-	void somm(int x, int y) {
-		int result;
+	void somm(float x, float y) {
+		float result;
 		result=x+y;
 		System.out.println("somma: "+result);
 	}
 	
-	void div(int x, int y) {
+	void div(float x, float y) {
 		float result;
 		float x1=x;
 		float y1=y;
